agents/app/agents/checkpointer.py
"""MongoDB checkpointer setup for LangGraph agent state persistence (task 4.13).

Uses ``pymongo`` directly to create a MongoDB client and wraps it in a simple
checkpointer that LangGraph can use.  The heavy ``langgraph-checkpoint-mongodb``
package is listed as an optional extra so tests can run without it installed.

When ``langgraph-checkpoint-mongodb`` is not installed, we fall back to a thin
in-memory shim so that unit tests (which mock MongoClient anyway) still pass.
"""

from __future__ import annotations

import os
from dataclasses import dataclass, field
from typing import Any

try:
    from pymongo import MongoClient  # type: ignore[import]
except ImportError:  # pragma: no cover
    MongoClient = None  # type: ignore[assignment,misc]


_DEFAULT_MONGO_URI = "mongodb://localhost:27017"
_DEFAULT_DB_NAME = "legal_ai"
_DEFAULT_COLLECTION = "checkpoints"


@dataclass
class CheckpointerConfig:
    """Configuration for the MongoDB checkpointer."""

    mongo_uri: str = field(default_factory=lambda: os.getenv("MONGO_URI", _DEFAULT_MONGO_URI))
    db_name: str = _DEFAULT_DB_NAME
    collection_name: str = _DEFAULT_COLLECTION


class _SimpleCheckpointer:
    """Minimal in-memory/MongoDB checkpointer shim.

    Provides ``put`` / ``get`` methods compatible with LangGraph's checkpointer
    interface.  When ``langgraph-checkpoint-mongodb`` is available it should be
    used instead; this shim exists so that the rest of the application can
    import and instantiate the checkpointer without installing optional extras.
    """

    def __init__(self, collection: Any) -> None:
        self._collection = collection
        self._memory: dict = {}

    def put(self, thread_id: str, checkpoint: dict) -> None:
        """Persist a checkpoint for *thread_id*."""
        self._memory[thread_id] = checkpoint
        if self._collection is not None:
            try:
                self._collection.replace_one(
                    {"thread_id": thread_id},
                    {"thread_id": thread_id, "checkpoint": checkpoint},
                    upsert=True,
                )
            except Exception:  # pragma: no cover
                pass  # Gracefully degrade if MongoDB is unavailable

    def get(self, thread_id: str) -> dict | None:
        """Retrieve the latest checkpoint for *thread_id*."""
        if thread_id in self._memory:
            return self._memory[thread_id]
        if self._collection is not None:
            try:
                doc = self._collection.find_one({"thread_id": thread_id})
                if doc:
                    return doc.get("checkpoint")
            except Exception:  # pragma: no cover
                pass
        return None

    def aget(self, thread_id: str) -> dict | None:
        """Async-compatible alias for :meth:`get`."""
        return self.get(thread_id)


def build_mongodb_checkpointer(
    config: CheckpointerConfig | None = None,
) -> _SimpleCheckpointer:
    """Build and return a MongoDB-backed LangGraph checkpointer.

    Parameters
    ----------
    config:
        :class:`CheckpointerConfig` instance.  If *None*, a default config
        reading ``MONGO_URI`` from the environment is used.

    Returns
    -------
    _SimpleCheckpointer
        A checkpointer instance backed by the specified MongoDB collection.
    """
    if config is None:
        config = CheckpointerConfig()

    if MongoClient is None:  # pragma: no cover
        raise ImportError(
            "pymongo is not installed. Run: uv sync --extra agents"
        )

    client = MongoClient(config.mongo_uri)
    collection = client[config.db_name][config.collection_name]

    return _SimpleCheckpointer(collection=collection)
