agents/pyproject.toml
[project]
name = "legal-ai-tool-agents"
version = "0.1.0"
description = "Legal AI Tool — Python agent backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    # Phase 4: LLM Gateway & agents
    "anthropic>=0.40.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    # Phase 4: PII
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    # Phase 4: SSE
    "sse-starlette>=2.0.0",
    # Phase 4: JWT
    "python-jose[cryptography]>=3.3.0",
]

# Production ML/RAG dependencies (heavy — install separately in production):
#   uv sync --extra ml
# Combines all extras:
#   uv sync --extra all
[project.optional-dependencies]
ml = [
    # Embeddings: pulls in PyTorch (~500 MB on macOS arm64)
    "sentence-transformers>=3.0.0",
]
rag = [
    "pinecone>=5.0.0",
    "llama-parse>=0.5.0",
]
agents = [
    # LangGraph MongoDB checkpointer (heavy — requires running MongoDB)
    "langgraph-checkpoint-mongodb>=0.0.4",
    "pymongo>=4.6.0",
    # bge-reranker (requires PyTorch)
    "FlagEmbedding>=1.2.0",
]
all = [
    "legal-ai-tool-agents[ml,rag,agents]",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
